requirements.txt
flask
flask-cors
google-cloud-storage
tensorflow
Pillow
